+++ ogma-cli/src/CLI/CommandFPrimeApp.hs
@@ -105,7 +105,7 @@ commandOptsParser = CommandOpts
             )
         )
   <*> strOption
-        (  long "app-target-dir"
+        (  long "target-dir"
         <> metavar "DIR"
         <> showDefault
         <> value "fprime"
@@ -113,7 +113,7 @@ commandOptsParser = CommandOpts
         )
   <*> optional
         ( strOption
-            (  long "app-template-dir"
+            (  long "template-dir"
             <> metavar "DIR"
             <> help strFPrimeAppTemplateDirArgDesc
             )