+++ ogma-cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Revision history for ogma-cli
 
+## [1.12.0] - 2026-01-21
+
+* Version bump 1.12.0 (#336).
+* Add CI job to test diagram backend (#332).
+* Standardize argument names across commands (#334).
+
 ## [1.11.0] - 2025-11-21
 
 * Version bump 1.11.0 (#325).