+++ ogma-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Revision history for ogma-core
 
+## [1.12.0] - 2026-01-21
+
+* Version bump 1.12.0 (#336).
+* Add support for state machine mermaid diagrams (#328).
+* Add support for sequence diagrams in Mermaid (#330).
+
 ## [1.11.0] - 2025-11-21
 
 * Version bump 1.11.0 (#325).